api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
psycopg2-binary==2.9.9
paho-mqtt==2.1.0
pydantic==2.9.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
